MGMT-1668: Automatic port allocation per namespace
We want to be able to run multiple namespaces on same hosts.
Currently it is possible, but the external ports of the uis
and inventories of each namespace overrides the connection
to the same port (6000 for inventory, 6008 to ui).

In order to be able to expose all ports, added bash logic
that:
1. checks for the next free port for a service, by a given
start port.
2. xinetd files and services names will be unique also by
namespaces, so services from different namespaces won't
override each other.
3. deletes previews allocated port per namespace (if exist).
In case we deploy same namespace twice, we want to reuse the
same port.