MGMT-1667: Use minikube profile per namespace
Currently Minikube VM starts with 8 GB of memory.
If we want to run two namespaces we need more then 8 GB.
When the VM will run out of memory the pods will be set in
pending state. We will have to destroy the running cluster
and start Minikube again with more memory.
We should be able to create or destroy a new namespace
without effecting the existing one using profiles.
The profile will be the namespace value, each namespace will
be created in its own VM.
This way it will be easier to manage the Minikube clusters.