[MGMT-766][MGMT-1108] Provide HTTP/S Proxy for OCP cluster installation
The environment variables HTTP_PROXY_URL / HTTPS_PROXY_URL could be
provided in order to set the Cluster Proxy on the installed cluster.

The HTTP Proxy will be used as a replacement of the deprecated PROXY_URL
variable for communication of the agent with the assisted-installer service.

Depends on openshift/assisted-service#41

Signed-off-by: Moti Asayag <masayag@redhat.com>