MGMT-1765: Execute oc commands from skipper
In order to deploy nodes against remote servers we need to be able
to execute oc commands from skipper, e.g perform oc login.
It does not enough to mount the kubeconfig file, we need to perform
the login actively.

This commit offers two ways to login. The simple one is to execute
the login command locally. Then the oc_login script will build the
command from the mounted kubeconfig file and perform the login.
The second option is to pass the oc command parameters manually.