Handle assisted-service user auth (#71)
* Disable assisted-service auth by default

This will prevent assisted-service from trying to authenticate
user api requests against a cloud.redhat.com account

* Add support for passing an offline auth token for assisted-service API auth

This patch uses the token to request jwt tokens from sso.redhat.com
A new jwt token will be requested each time it expires

* Use SSO_URL env to override redhat sso endpoint

* Check for empty SSO_URL env var

When skipper substitutes env vars it will create one in the
target environmentif you don't have a value set. This leads
to an empty string for SSO_URL when we don't want to override

* Check for empty OFFLINE_TOKEN

This is needed because skipper will set an env var to an empty
string if it's in skipper.yaml and unset in the calling environment

* Move SSO_URL default to Makefile

If it's not set in assisted_service_api.py we error out with a key
error.