Raz/mgmt 1668 pord forwarding by namespace (#106)
* mgmt-1668 - added namespace indexer

In order to determine which port to allocate (in the
future also cidr and network bridge) we will set a
unique index for each namespace. The ports for the ui
and assisted service will be the base ports + index.

* mgmt-1668 - allocate port using NAMESPACE_INDEX

Instead of using hard coded ports. Port name is the
base port + NAMESPACE_INDEX. Open firewalld port on
the port spawning sections.

* mgmt-1668 - updated makefile calls with ns indexes

Added calls to allocate and retrieve index per
namespace on service deployment commands. Also updated
relevant port kill commands.