Raz/mgmt 1763 multi terraform env (#115)
* MGMT-1763: refactor oc_utils.py - no logic change

In this commit I broke OC class for cofiguration and client classes
that are more logically correct and will serve us in next commit
when we want to use kubernetes client.

* MGMT-1763: start_discovery.py create a unique env per instance

In order to allow multiple terraform environments on the same host,
default terraform variables changed to be unique for the cluster
name.

1. cluster name will include namespace postfix instead of random postfix.
2. iso file name, terraform dir and network name will include the
cluster name.
3. if not provided, automatically search for available cidr address,
network bridge and free ips for the nodes.

* MGMT-1763: install_cluster.py use new cluster name

Accept custom cluster name from user, or use the new convention
with the namespace postfix.

* MGMT-1763: delete_nodes.py delete nodes per environment

1. use new cluster naming convention.
2. if namespace not provided, try to delete all namespaces.
3. added destroy_all_nodes target to makefile: delete all nodes without deleting the assisted-installer local deployment.

* MGMT-1763: add random postfix to ui cluster name

In order to avoid conflicts with previews/other clusters.

* MGMT-1763: update delete_all_virsh_resources

it will try first to delete all namesapaces under minikube
profile. If failes, tries under the <namespace> profile.

* MGMT-1763: remove error on del index when namespace not exist

Since the namespace may be created not using
assisted-test-infra, yet we don't want this error to break
the deletion flow.

* MGMT-1763: allow only one instance of run terraform

Since it might get errors if two instances are starting in parallel.

* MGMT-1763: updated Makefile and Jenkinsfile

* MGMT-1763: Jenkinsfile - profile will be equal to namespace

It should not be exposed as parameter, instead it should get
the namespace value.