NO-ISSUE: Added KubeAPIContext
KubeAPIContext is an object which holds information about the resources
created during a test. It will be used in a fixture
called kube_api_context and provides the kube_api_client. By doing that,
there will be no need to worry about cleanup of any resource that will be
created in the test (i.e no need to run the test within contextmanager).