MGMT-5274 Deploy assisted operator
Add new 'operator' DEPLOY_TARGET to deploy_assisted_service.sh that
calls deploy/operator/deploy.sh script from assisted-service repository.

The current workflow to install an operator is to make a regular
installation with AI, and afterwards use its KUBECONFIG to deploy the
assisted-operator on top of it.

A cluster installed with test-infra is provisoned with nat networking.
It created iptables rules by design to ignore inbound connections except
from the host.
Hence, in order to make new VMs from a new network to reach the
assisted operator that is installed on the cluster - we have to pass the
communication through the host. It is made by overriding
SERVICE_BASE_URL.

Before deploying, we will start an nginx load balacer and configure it
to listen to http and proxy pass the communication to the OCP route.
The OCP route looks at the layer 7 Host field for destination - so we
change that as well with "proxy_set_header Host" configuration.