MGMT-6545: fix operator resources (#886)
* Get all supported OLM operators from the service for test_olm_operator

* MGMT-6545: fix operator resources

This PR introduces a fixture to update cluster and terraform configuration
according to requirements needed per specific operator. It is very misleading
that the number of nodes ae in both configs and it is not enough to update
`workers_count` but we need to do it for duplicated value of `nodes_count`.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>

Co-authored-by: Yuval Goldberg <yuvigoldi@gmail.com>