MGMT-7523: Duplicated `get_nodes` fixture into `get_nodes_infraenv`
The only difference being is that the TerraformController receives an InfraEnvConfig
instead of a ClusterConfig. This in-turn causes it to launch nodes using
our `baremetal_infra_env` config dir rather than our baremetal config dir.
The difference between those being that the latter does not create cluster DNS configuration which is unnecessary when you don't have a cluster