MGMT-7523 - Added initial kube-api late binding test (#1098)
* MGMT-7523: Added pyproject.toml with tool.pyright `extraPaths` directive

discovery-infra is assumed by the test-infra code as a place to look for libraries that are imported. This tells pyright about it being that, which causes pyright to work properly in this project

* MGMT-7523: Added missing status types to the AgentStatus enum

* MGMT-7523: Fixed `NodeController.get_machine_cidr` return type hints

* MGMT-7523: Fixed formatting of `ClusterName` and `InfraEnvName`, added explicit kwargs

* MGMT-7523: Fixed creation of spoke cluster namespace to be idempotent

* MGMT-7523: `get_default_openshift_version` added some logs / more verbose errors to clarify assertion

* MGMT-7523: Fix `download_iso_from_infra_env` unnecessary dependency on cluster_config to get the download target path

This allows it to be used in the late-binding flow where we don't have a cluster_config

* MGMT-7523: get_nodes fixture type hint was wrongly too generic

* MGMT-7523: InfraEnv.create now accepts a None cluster_deployment to allow late-binding

* MGMT-7523: Split off some of `wait_for_agents_to_install` into `wait_for_agents_to_be_ready_for_intall`

When creating an infra-env without a cluster-ref, we care to check if the
agents are ready for install, so we can know when we can bind them to
a cluster

* MGMT-7523: Added type hint that the IDE was having a hard time to infer

* MGMT-7523: Split `configs` fixture, added infraenv_config fixture

This is to facilitate late binding tests which need just a terraform_config and an infraenv_config

* MGMT-7523: Added `wait_for_agents` and `list_agents` to infra-env

The regular kube-api tests just listed the cluster agents by looking
at all agents and filtering according to their clusterRef. Now agents
can be created without a cluster, where they only belong to an infraenv.

This allows listing the agents belonging to a particular infraenv. The
filtering is done by judging the agents' `infraenvs.agent-install.openshift.io`
label.

* MGMT-7523: Duplicated `get_nodes` fixture into `get_nodes_infraenv`

The only difference being is that the TerraformController receives an InfraEnvConfig
instead of a ClusterConfig. This in-turn causes it to launch nodes using
our `baremetal_infra_env` config dir rather than our baremetal config dir.
The difference between those being that the latter does not create cluster DNS configuration which is unnecessary when you don't have a cluster

* MGMT-7523: Added `kube_test_configs_late_binding` fixture

This fixture is similar to `kube_test_configs` except that is created
an infraenv_config fixture rather than a cluster_config

* MGMT-7523: Created `test_kube_api_late_binding_ipv4` test

For now all the test does is create an infra env and boot agents into it.

In the future, it will be expanded to also create a cluster and bind the
agents to that cluster.

* MGMT-7523 - Change all instances of Union[<>, None] to Optional[<>]