MGMT-7902: Test infra IPv6 static ips not working (#1135)
When using test infra, in order to match the hostname the actual host in
the cluster, test-infra is using interfaceAddresses method on libvirt
domain object, which provides the macs and ips for the domain (VM).
Libirt supports 2 methods to gather this information: DHCP and ARP.  ARP
only exists for IPv4, and no DHCP information exists when using static
ips.
To retrieve this information, the tfstate file is parsed.  It already
contains the hostname and the mac addresses of the interfaces.  By
matching the mac addresses in the inventory, the hostname and roles can
be set correctly for this configuration.