MGMT-7957 - `test_kube_api` fix to use newer `ClusterConfig` members
aec453f5fcf1da37e6171940df71fef1bc96cf66 changed the format of the `ClusterConfig`
class but forgot to update all usages of it.

For now this is a quick fix just to get the kube-api CI working again, it is
not an entirely correct fix because it has poor support for dual network stack
clusters.

This may be improved in future PRs