MGMT-8300: Apply dual-stack configuration in a single step
With https://github.com/openshift/assisted-service/pull/3042 we are
making it possible to create a dual-stack cluster in a single API call.
At the same time, it's no longer possible to create a dual-stack cluster
without providing Machine Networks from the very beginning. Because of
that, we are adapting test-infra to send the full cluster payload
instead of splitting it into two separate steps.

Contributes-to: MGMT-8300