MGMT-8660: Fix os image override script
Previously this script would edit the image entry in place as well as
appending a new image to the list.

The existing entry should not be edited and the new entry should be made
unique by manufacturing a "version" value. This will prevent the image
service from failing due to MGMT-8661.

https://issues.redhat.com/browse/MGMT-8660
https://issues.redhat.com/browse/MGMT-8661
https://bugzilla.redhat.com/show_bug.cgi?id=2012844