Archive must-gather directory (#1507)
Following "KAS as a service", it became evident that it's better storing
must-gather logs as a tar archive, than leaving it in a directory. This
not only saves data, but also something that is more common among other
jobs. In addition, downloading the directory to the computer takes more
time as it has to be done recursively (and GCS is rate-limitting it
per-file)