Use pypi's assisted-service-client
Now that the assisted-service-client is published in the Python Package
Index, it makes sense to use that instead of copying it from the
assisted-service image.

Additionally, this should make it easier once we branch out to get the
matching assisted-service-client for specific branches rather than the
latest upstream. Though the assisted-service-client should be backwards
compatible, so it is not a big deal

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>