NO-ISSUE: Increase libvirt wait for machine IP timeout from 360 to 600
In this run:
https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-assisted-service-master-e2e-metal-assisted-kube-api-late-binding-single-node-periodic/1503164013213650944

It seems like this function timed-out but the machine did eventually
get an IP address. Hopefully this timeout increase will prevent such
scenario from occurring in the future.

This issue seems to be common for this job.

It's also possible there's an actual bug and this will not really fix
this issue, time will tell.