NO-ISSUE: [release-ocm-2.4]: Change catalog image to one that's actually being updated
This is actuallyr related to https://bugzilla.redhat.com/show_bug.cgi?id=2072892
but I put NO-ISSUE because I don't want the bz bot involved (as there's
another PR openshift/assisted-service#3635 managing that bug).

`INDEX_IMAGE` was set to
`quay.io/ocpmetal/assisted-service-index:latest` which is wrong because
we actually need `ocm-2.4` and not `latest`, and also the organization
`ocpmetal` is no longer being published to, it should be
`edge-infrastructure` instead.