Handle forward-compatibility with 'setup' target (#1606)
Back in #1442, we've had changed our makefile targets to remove some of
the redundancy. Most important one to handle is the
``create_full_environment`` -> ``setup`` change, because ``create_full_environment``
is heavily referenced in CI.

This PR introduce ``setup`` target so that we can use it in
cross-branches places, e.g. assisted-baremetal-setup step.