MGMT-8938: Add support for bootstrap-in-place day-2 worker flow (#1639)
Note: These changes are for tests of the BIP flow in this repo
and they don't have anything to do with the Assisted Installer.

# Description

This commit adds a new Makefile target `deploy_ibip_worker` which runs a
new pytest test called `test_bip_add_worker`.

This new pytest "test" is identical to the existing
`test_bootstrap_in_place` test except once the installation of the BIP
SNO is done, it also adds a worker to that SNO.

Once the test starts the worker, it also waits for its CSRs to show up
and then approves them.

This test is made possible due to 4.11 support of adding workers to SNO.
This new test will not work on <=4.10.

# Background

Note that the new worker shouldn't boot a live-ISO with the
installer-generated `worker.ign` directly. Instead, it has to boot a
blank RHCOS Live ISO, then manually run `coreos-installer install...`
with the `worker.ign` file to install RHCOS with the `worker.ign`.

To achieve this in this test, we create *another* ignition file out
of a template in `src/assisted_test_infra/resources/sno-worker-live.ign.j2`,
and in this ignition file we:

- Copy the `worker.ign` file into the live filesystem
- Create a one-shot systemd unit service which runs coreos-installer with that worker.ign
- Reboot the node once RHCOS has been installed to disk

The worker then starts with the installed RHCOS+worker.ign and attempts to join the cluster

# Notable changes

- Added a new Makefile target to run the new flow
- Made the "running" terraform variable configurable. This allows me to
  create both the worker and the SNO and then only later control when they start.
- Added new Terraform config variable called `worker_image_path` which
  controls the ISO image used for workers. By default, if unset, it
  defaults to using the same image as the masters (this is the behavior
  all users of the terraform controller depend on).