MGMT-9378 - Bug fixes to make day-2 Assisted REST flow support SNO (#1694)
# Background

Starting OCP 4.11 there is a requirement to support day-2 installation
of worker nodes on SNO clusters as opposed to just multi-node clusters.
We want Assisted REST to support it as well, so we want to run periodic
tests to make sure it works.

The existing Day2Cluster test flow first installs OpenShift using the
REST assisted-installer and then attempts to add worker nodes to the
resulting cluster. It worked fine in multi-node clusters but had some
bugs for single-node clusters (NUM_MASTERS=1 and NUM_WORKERS=0) which
this commit aims to solve.

# DNS Fix

The first bug is the missing `api-int.` DNS entry in the terraform
libvirt dnsmasq configuration. Only the `api.` entry was present. Since
multi-node clusters use the baremetal platform rather than the "none"
platform, this wasn't a problem for them as DNS is taken care of in that
platform automatically. But single-node clusters use the "none" platform
and so require DNS to be manually configured. This has been solved by
enabling the `api-int.` DNS configuration in terraform.

# `are_libvirt_nodes_in_cluster_hosts` Fix

The second bug was trouble with the `are_libvirt_nodes_in_cluster_hosts`
implementation. It was, for some reason, failing to run on the Assisted
cluster when its in "adding-hosts" mode. Not really sure why because the
previous implementation hid the exception, but I've reimplemented it
without the unnecessary inventory/mac stuff that wasn't really used
anyway, and it works fine now.