MGMT-10913: Use our fork of dmacvicar/terraform-provider-libvirt
https://github.com/dmacvicar/terraform-provider-libvirt is recently
unmaintained and doesn't seem to be merging pull requests.

There are some changes we wanted to make in the project but couldn't
because the maintainer haven't merged our pull requests for a while now,
so we decided we should [fork it](https://github.com/openshift-assisted/terraform-provider-libvirt) instead for now.

Changes include:
- Race condition when creating two libvirt networks at the same time
- Some missing ARM support (@tsorya can elaborate)
- Like three other bugs / issues mentioned in our code if you grep for dmacvicar

This pull request makes it so that when running tests, the fork will be
cloned, built, and installed. All terraform files have been modified
to use the fork.