NO-ISSUE: Remove Fedora support, add warning about this repo (#1708)
Users are not recommended to run the code in this repo on their local
machine, as it makes some surprising opinionated configuration changes
that users might not expect (installing packages, configuring libvirt,
etc). We should give a clear warning about that in the README.

Also since we don't support Fedora, our `create_full_environment.sh`
should be modified to reflect that and block users trying to install
on Fedora. If we detect an OS we don't support, on top of saying we
don't support it, we give the same warning as in the README.