MGMT-10775: Use smaller IPv4 range for libvirt DHCP
This change modifies the DHCP range used by the Terraform-created
libvirt network. This is to mitigate an issue where API and Ingress VIPs
that we are statically defining happen to be assigned via DHCP to the
VM. This creates a failure with the following error

```
ingress-vip <192.168.127.101> is already in use in cidr 192.168.127.0/24
```

With this change the DHCP range is shortened to cover addresses only up
to `.90` what is enough for our testing purposes as we will not be
creating so many VMs in the foreseeable future.

Closes: MGMT-10775