NO-ISSUE: allow cross network traffic
The rootfs is now downloaded from the assisted-image-service instead of
internet.

As a consequence, when we try to setup a cluster located out of the hub
cluster network, the machine that boots the assisted ISO fails to
retrieve the rootfs because libvirt denies the traffic between guest
networks.

This change sets up a hook in libvirt that drops the rules denying
forwarded traffic between guest networks.

Here is a job failing because of this issue:
https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-assisted-service-master-edge-e2e-metal-assisted-kube-api-late-binding-single-node-periodic/1544506948760637440