Fix swtpm installation not being idempotent (#1775)
When doing ``make setup`` twice, it fails on the ``swtpm`` dir already
existing:
```
Install swtpm v0.7.1 from source
fatal: destination path 'swtpm' already exists and is not an empty
directory.
make: *** [Makefile:128: setup] Error 128
```

In addition, it leaves unstaged files in the repo that are not ignored
by git.

This change will use a temporary dir from ``/tmp``, and will cleanup
afterwards with a bash trap.