NO-ISSUE: Fix late-binding job (#1864)
Two issues:

1/
Since #1852 the
loadbalancer configuration is cleaned up before starting it.

This change does not cleanup the configuration when TEST_TEARDOWN is
set to false.

It should fix the late binding job, that configures the nginx server
before running the tests.

2/
Since https://github.com/openshift/assisted-test-infra/pull/1849/files#diff-21e688d6eb3b87727264f6934118765f035f5bf7504dff17b3ab9fb388f4d054L416

The kubeconfig used for the hub cluster is replaced by the kubeconfig that points
to the sno cluster which leads to fail the conformance test as the DNS for this cluster
is not configured.

I made the parameter optional, as I guess the kubeconfig from the AgentClusterInstall
is not used in this case.