Fix DNS for kube-api SNO installations
The kube-api-net-suite job fails running conformance tests because of
invalid DNS values that are configured in the system:
```
Get
"https://api.test-infra-cluster-c418d125.redhat.com:6443/version?timeout=32s":
dial tcp 192.168.128.100:6443: connect: no route to host
```

(see
https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/33427/rehearse-33427-pull-ci-openshift-assisted-service-master-edge-e2e-metal-assisted-kube-api-net-suite/1586727810129465344)

The reason is that we pick ``<subnet>.100`` for the access IPs (API and
ingress) in the case of single-node clusters, where it should have
picked this node's IP instead.

This change refactors a bit of the code that uses ``ipaddress`` module
for readability, and removes the option to use
``is_highly_available=False`` option, (as it's mostly unused and only
confusing). I checked and QE's tests are not using this option, so it
should be safe to break this functionality.