MGMT-8707: Revert swtpm installation from source (#1916)
Manual revert of
https://github.com/openshift/assisted-test-infra/pull/1748.

The fixed versions of swtpm packages has landed for RHEL 8.7+8.6 as well
as latest Rocky Linux 8 version.

It means we should not have any kind of flaky behavior as we previously
had in CI (might still happen locally, but either way disk-encryption is
not heavily prioritized for testing and development).