Allow overriding pytest flags
Back in https://github.com/openshift/assisted-test-infra/pull/1824 we
prevented tests marked with skip to just be skipped (currently they are
marked as failures).
The reason we did it is to prevent a situation where the list of
openshift versions come off empty and the tests are considered skipped
(and passing vacuously).

This enables emptying all pytest ``PYTEST_FLAGS`` to disable this kind
of behavior.