MGMT-13210: only pull relevant OS images
Currently we load all images from data/default_os_images.json in prow
CI. It makes for a very long setup time for the image-service.

At least some of those images are not relevant to our CI at all, because
we don't have a way to get the matching HW - namely the images for
zSystems and PowerPC.

In addition, we don't need to load images for 4.8, 4.9, 4.10, etc. for
most of the jobs as we rather use 4.11/4.12. So it should be beneficial
to only load the relevant image based on existence of OPENSHIFT_VERSION
or OPENSHIFT_INSTALL_RELEASE_IMAGE env-vars.