Pulling internal registry from quay.io (#1988)
In local environments, we sometimes hit the rate-limitter of docker hub
when doing a lot of ``make destroy run``. This is because the registry
addon of minikube relies on images that are not in ghcr.io but from
docker-hub.

This change uses images from quay.io for the registry, as a simple fix.