MGMT-13341: Use assisted service client wheel (#2005)
Currently, the sdist version fails to install because of an update of
setuptools: https://github.com/pypa/setuptools/issues/3772.

The wheel is not affected by this change.

Depending on the output of the setuptools ticket, we may have to
follow-up to provide a valid version scheme for the sdist package.

Co-authored-by: Adrien Gentil <agentil@redhat.com>