NO-ISSUE: update libvirt hook to be lighter (#2036)
At some point of time, my machine was consuming 100% CPU on
iptables-restore/iptables-save and I had to reboot the machine to
recover it.

This change deletes only the rules that need to be deleted instead of
doing a full save/restore of all iptables rules. I expect this process
to be lighter.