MGMT-12442: Sync iso images between heterogeneous hosts
Create a shared mount between the heterogenous hosts, so when the ISO
image is downloaded on the primary, the sedondary (ARM machine) hosts as
also access to it automatically.

This chnages creates a new role to setup the sshfs mount, fixes some
linter issues in the ssh role and lower the ansible log level.