Add hub cluster k8 client wrapper to allow upgrade and scaleup (#2053)
Added a base code for hub cluster controller and a k8_minikube fixture.

minikube as hub cluster  initialize instance with k8_client and allow
access to resources and change configuration, scaleup/down simulate
stop and start. in addition the code allows to change confiugration to any
resource object plus waiters till stop/start completed.

Each hub cluster resource has a kind like ConfigMap, Pod .. we get a reference
to the kind type and look for the instance object base on lookup name.
Support for change configuration per resource kind type and verify that
configuration really changed.

scaleup/down simulate stop/start . when stop we have a waiter and verify
that Development and ReplicaSet counters set and till the pods are deleted.
When start we verify that Development and ReplicaSet counter/ status set
and Pod are up with valid status and expected number of replicas (default 1)

Added to BaseTest k8s_assisted_installer_infra_controller fixture
Create object for infra controller as minikube and store ConfigMap configuration
We verify that all Developments are up and running.
Cleanup fixture happens when test done, we compare ConfigMap after and
restore configuration only when there was a patch for change configuration.
cleanup has waiter in case configmap changed , we verify that configuration
restored for the next testcase.

The test should run in serila mode only.

Planning to add more capablities to the minikube_controller
- Adding metrics for hub cluster configuration and track kubectl top pods for cpu and mem