Day2 post installation triggers node reboot while addding csr (#2080)
Current flow for day2 installation try to approve workers csr while the node
is rebooting and we keep trying in the waiter.
Added sshable waiter after day2 installation to verify that node really returns from
reboot and approve the csr when node is up and running.

We have failures in kni-auto with day2 and it might be a timeing issue depends on reboot
time of the installed node.