MGMT-14001: Generate events.html as standalone file
Due to the nature of Jira attachments, it's not possible to dynamically
load JSON event files from the events.html based on relative placement
of the files.

We can make the events.html file a standalone file (one that you can
download by itself and will still show all relevant data) to cope with
that issue.