Do not download kubeconfig if must-gather wasn't requested
Currently test-infra tries to download ``kubeconfig-ingress`` even if
we're not using it for gathering must-gather logs.

In triage logs download process we're having this kind of behavior, even
though it doesn't succeed due to roles scheme. It produces a warning
like:
```
18:23:56  2023-03-21 16:23:55,895  download_logs WARNING    -
   140242975606592 - Suppressed ApiException from download_logs:246 :
		   (403)
18:23:56  Reason: Forbidden
18:23:56  HTTP response headers: ...
18:23:56  HTTP response body:
   b'{"code":403,"message":"<user>: Unauthorized to access
route (insufficient role read-only-admin)"}'
```

This only does the download if configured to try and gather must-gather,
making it produce less warning logs.