Fix stale references to infra-envs (#2101)
Currently the method for fetching infra-env objects is as follows:
* Listing all existing clusters (including ones that are logically
  removed)
* Fetching cluster object from API (mainly to get the associated hosts
  which are not retrived in the first stage)
* Fetching infra-env objects from API based on IDs in the host objects

The problem is step 3, as infra-env objects might get deleted in the
meantime, meaning those IDs no longer reference to existing infra-envs.

To minimize the problem, this change ignores infra-env IDs in the host
objects, and just do another request for the infra-envs that are
associated with a certain cluster ID.
It guarantees a more recent representation of current DB state, but more
importantly we shouldn't get 404 for requesting those objects.
Later on we download events for those infra-envs, but this action is
suppressed in case of API errors.