Add support for multiple assisted-service pods
Added support for:
- restart_deployment
We can restart the deployment without stop , the restarted managed and creates
another replicas and manages the pods creation/termination. will allow us to run
tests during agent upgrades as active/active
- delete_leftovers , due to the way restartAt works we have to clean all leftovers
for the next run, before cleanup we remove replicaset with zero or not running pods
- containers_ready_inside_pod
Added support for deep inspection of a pod and verify that inside containers are
really marked as ready. also we calculate the total pods plus verify that each one
is in ready (from conainter status)

At the end of the test we back to the last good state with a single pod, prefer to do it with
oc scale replicas because there are no replicas leftover