Update old image used by upgrade agent test (#2171)
The upgrade agent prepares a cluster for installation with an old agent
image and then checks that it can be upgraded to the latest one. But
that old image is no longer capable to move the cluster to ready because
of changes to the DNS resolution validation. To address that issue this
patch changes the test to use a newer old image.

Related: https://issues.redhat.com//browse/MGMT-14526
Related: https://github.com/openshift/assisted-service/pull/5208
Related: https://github.com/openshift/assisted-installer-agent/pull/542

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>