Increase pending user input timeout to 45 minutes (#2192)
When installing a cluster we may hit on pending user input when wrong
boot order detected .
We should take care of two sceanrios
- Master node is not a bootstrap and got rebooted after write to disk phase
- Master bootstrap node , rebooted only when other master's installed.

In case bootstrap node the timeout is not enough.

We can address it in other ways also like use kwargs in helper function or
fix a private test case if needed.