Add support for day2 master and mixed worker and masters (#2181)
Current code add support for day2 master nodes and allows to increase
number of masters in day2 or workers.

Added day2_masters_count.
Added set_hostnames_and_roles in day2_cluster to set the proper role for node

Code tested in kni and was able to install worker and master as day2
Current code approved CSR for the nodes in ocp and the nodes status was READY.

Next phase may be to add additional command to verify day two installation