Fix pip incompatibility with Python version
It seems like ``pip`` is using functions that are unavailable in the
Python distribution that wer'e using. To try and mitigate the issue,
we'll first check if we can opt to use a newer Python version (chosen
3.11 as it's the latest stable version that we're also using for the
container image) and fallback to the older packages if none could be
found.